Changed MultiplyFunctor -functor class to use float values, was using doubles that were causing compiler warnings.

+++ libs/fmi/bizcode/tm_utils/source/ToolMasterHelperFunctions.cpp
@@ -21,12 +21,12 @@ namespace
     // k‰ytet‰‰n avuksi least-squeare-metodin yhteydess‰
     struct MultiplyFunctor
     {
-        MultiplyFunctor(double theFactor) :itsFactor(theFactor) {}
-        double operator()(double theValue)
+        MultiplyFunctor(float theFactor) :itsFactor(theFactor) {}
+        float operator()(float theValue)
         {
             return theValue*itsFactor;
         }
-        double itsFactor;
+        float itsFactor;
     };
 }
